worker/requirements.txt
celery==5.3.6
sqlalchemy==2.0.19
psycopg2-binary==2.9.9
ffmpeg-python==0.2.0
python-dotenv==1.0.1
pika==1.3.2
requests==2.31.0
